# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dhruva PGRS API
  version: 1.0.0
  description: |
    Andhra Pradesh Public Grievance Redressal System (PGRS) API

    **Core Innovation**: Verification-First Architecture
    - Citizens MUST approve before case closure
    - Immutable audit trail (tamper-proof)
    - Multi-channel intake (web, mobile, WhatsApp, SMS, voice)

    **Philosophy**: Antifragile, Modular, Test-Driven Development

    ## API Versioning Policy

    **Current Version**: v1.0 (Stable)
    **Base URL**: `https://api.dhruva.ap.gov.in/v1`

    ### Version Support Lifecycle

    | Version | Status        | Launch Date | Deprecation Date | Sunset Date  |
    |---------|---------------|-------------|------------------|--------------|
    | v1.0    | Stable        | 2025-12-01  | TBD              | TBD          |
    | v2.0    | Planned       | 2026-06-01  | -                | -            |

    ### Support Phases

    1. **Stable** (Current)
       - Fully supported with bug fixes and security patches
       - Backward-compatible changes only
       - No breaking changes

    2. **Deprecated** (When v2.0 launches)
       - v1.0 remains operational for 12 months after v2.0 launch
       - Security patches only (no new features)
       - Response headers include deprecation warnings
       - Clients encouraged to migrate to v2.0

    3. **Sunset** (12 months after deprecation)
       - v1.0 API shut down
       - All requests return HTTP 410 Gone
...

    ### Deprecation Headers

...

...

...

...
servers:
...
paths:
...
  /api/v1/grievances/{grievance_id}:
    get:
      tags:
        - Grievances
      summary: Get grievance by ID
      description: |
        Get detailed information about a specific grievance.

        **Performance Budget**: p95 < 300ms
        **Authorization**: Officers can view assigned grievances, Admins can view all
      operationId: getGrievanceById
      security:
        - bearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          description: Grievance UUID or public ID (PGRS-YYYY-DD-NNNNN)
          schema:
            type: string
          examples:
            uuid:
              value: "550e8400-e29b-41d4-a716-446655440000"
            public_id:
              value: "PGRS-2025-05-00001"
      responses:
        '200':
          description: Grievance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrievanceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    patch:
      tags:
        - Grievances
      summary: Update grievance
      description: |
        Update grievance status, assignment, or resolution.

        **Performance Budget**: p95 < 500ms
        **Authorization**: Officers can update assigned grievances, Supervisors can reassign
        **Audit Trail**: All updates logged in audit_logs table
      operationId: updateGrievance
      security:
        - bearerAuth: []
      parameters:
        - name: grievance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrievanceUpdateRequest'
            example:
              status: "in_progress"
              resolution_notes: "Inspected site. Scheduling repair work."
      responses:
        '200':
          description: Grievance updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrievanceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - status_code
      properties:
        error:
          type: string
          description: Error type (e.g., "Validation Error", "Unauthorized")
          example: "Validation Error"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid phone number format"
        status_code:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: object
          description: Additional error details (validation errors, etc.)
          additionalProperties: true
        retry_after:
          type: integer
          description: Seconds until retry allowed (for rate limits)
          example: 60
    ValidationErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Field name that failed validation
          example: "citizen_phone"
        message:
          type: string
          description: Validation error message
          example: "Invalid phone format. Use +91XXXXXXXXXX"
        code:
          type: string
          description: Machine-readable error code
          example: "INVALID_PHONE_FORMAT"

    # ==========================================
    # AUTHENTICATION SCHEMAS
    # ==========================================
...
    UserResponse:
      type: object
      required:
        - id
        - username
        - full_name
        - role
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        username:
          type: string
          example: "officer_vijayawada"
        email:
          type: string
          format: email
          example: "officer.vijayawada@dhruva.ap.gov.in"
        full_name:
          type: string
          example: "Ramesh Kumar"
        phone:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
          example: "+919876543211"
        role:
          type: string
          enum: [officer, supervisor, admin]
          example: "officer"
        department:
          $ref: '#/components/schemas/DepartmentResponse'
        district:
          $ref: '#/components/schemas/DistrictResponse'
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          nullable: true

    # ==========================================
    # GRIEVANCE SCHEMAS
    # ==========================================
...
    GrievanceUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [assigned, in_progress, resolved, verified, closed, rejected]
        assigned_officer_id:
          type: string
          format: uuid
          nullable: true
        department_id:
          type: integer
          nullable: true
          description: For manual department reassignment (supervisor only)
        resolution_text:
          type: string
          minLength: 20
          maxLength: 5000
          nullable: true
        resolution_notes:
          type: string
          maxLength: 2000
          nullable: true
    GrievanceResponse:
      type: object
      required:
        - id
        - grievance_id
        - citizen_name
        - citizen_phone
        - status
        - priority
        - district
        - sla_days
        - due_date
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        grievance_id:
          type: string
          pattern: '^PGRS-\d{4}-\d{2}-\d{5}$'
          example: "PGRS-2025-05-00001"
        citizen_name:
          type: string
          example: "Rajesh Kumar"
        citizen_phone:
          type: string
          example: "+919876543210"
        citizen_email:
          type: string
          nullable: true
        citizen_address:
          type: string
        status:
          type: string
          enum: [submitted, assigned, in_progress, resolved, verified, closed, rejected]
        priority:
          type: string
          enum: [critical, high, normal]
        department:
          $ref: '#/components/schemas/DepartmentResponse'
          nullable: true
        district:
          $ref: '#/components/schemas/DistrictResponse'
        assigned_officer:
          $ref: '#/components/schemas/UserResponse'
          nullable: true
        sla_days:
          type: integer
          example: 7
        due_date:
          type: string
          format: date-time
        grievance_text:
          type: string
        language:
          type: string
        channel:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        verified_at:
          type: string
          format: date-time
          nullable: true
    GrievanceDetailResponse:
      allOf:
        - $ref: '#/components/schemas/GrievanceResponse'
        - type: object
          properties:
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/AttachmentResponse'
            audit_logs:
              type: array
              items:
                $ref: '#/components/schemas/AuditLogResponse'
            verifications:
              type: array
              items:
                $ref: '#/components/schemas/VerificationResponse'
...
    DistrictResponse:
      type: object
      required:
        - id
        - code
        - name
      properties:
        id:
          type: integer
          example: 5
        code:
          type: string
          pattern: '^[0-9]{2}$'
          example: "05"
        name:
          type: string
          example: "Krishna"
    DepartmentResponse:
      type: object
      required:
        - id
        - code
        - name
        - sla_days
      properties:
        id:
          type: integer
          example: 2
        code:
          type: string
          example: "HLTH"
        name:
          type: string
          example: "Health Department"
        name_telugu:
          type: string
          nullable: true
          example: "ఆరోగ్య శాఖ"
        description:
          type: string
          nullable: true
        sla_days:
          type: integer
          description: Default SLA in days
          example: 7

    # ==========================================
    # ATTACHMENT, AUDIT, VERIFICATION SCHEMAS
    # ==========================================
    AttachmentResponse:
      type: object
      required:
        - id
        - file_name
        - file_type
        - file_url
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
        file_type:
          type: string
          enum: [image/jpeg, image/png, application/pdf, audio/mpeg, video/mp4]
        file_size_bytes:
          type: integer
        file_url:
          type: string
          format: uri
        attachment_type:
          type: string
          enum: [before, after, evidence, document]
          nullable: true
        uploaded_at:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      required:
        - id
        - action
        - created_at
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
          enum: [CREATED, UPDATED, ASSIGNED, RESOLVED, VERIFIED, CLOSED, REJECTED, ESCALATED, REOPENED]
        actor:
          $ref: '#/components/schemas/UserResponse'
          nullable: true
        old_value:
          type: object
          nullable: true
        new_value:
          type: object
          nullable: true
        reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    VerificationResponse:
      type: object
      required:
        - id
        - verification_method
        - message_sent_at
        - citizen_response
      properties:
        id:
          type: string
          format: uuid
        verification_method:
          type: string
          enum: [whatsapp, sms, call, email]
        phone_number:
          type: string
          nullable: true
        message_sent_at:
          type: string
          format: date-time
        citizen_response:
          type: string
          enum: [approved, rejected, pending]
        citizen_rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        response_received_at:
          type: string
          format: date-time
          nullable: true

    # ==========================================
    # OTP SCHEMAS
    # ==========================================
...
  responses:
    ValidationError:
      description: Validation error (invalid input)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      $ref: '#/components/schemas/ValidationErrorDetail'
          example:
            error: "Validation Error"
            message: "Invalid input data"
            status_code: 400
            details:
              - field: "citizen_phone"
                message: "Invalid phone format. Use +91XXXXXXXXXX"
                code: "INVALID_PHONE_FORMAT"
              - field: "grievance_text"
                message: "Minimum 20 characters required"
                code: "MIN_LENGTH_VIOLATION"
    UnauthorizedError:
      description: Unauthorized (invalid or missing token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Invalid or expired token"
            status_code: 401
    ForbiddenError:
      description: Forbidden (insufficient permissions)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Forbidden"
            message: "You do not have permission to access this resource"
            status_code: 403
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Not Found"
            message: "Grievance not found"
            status_code: 404
...